Renamed directory for handler
